requirements.txt
numpy
scipy
tensorflow-gpu
Pillow
matplotlib
jupyter
tqdm
scikit-learn==0.19.1
keras